basic authentication with jwt is implimented . jwt genarating authnetication autherisation and pasword hashing also iplimented

+++ src/main/java/com/shashimadushan/aliscapper/service/ProductService.java
@@ -1,30 +1,53 @@
 package com.shashimadushan.aliscapper.service;
 
+import com.shashimadushan.aliscapper.dto.ProductDTO;
 import com.shashimadushan.aliscapper.model.Product;
 import com.shashimadushan.aliscapper.repo.ProductRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class ProductService {
     @Autowired
     private ProductRepository productRepository;
+    private final ModelMapper modelMapper = new ModelMapper();
 
-    public List<Product> getAllProducts() {
-        return productRepository.findAll();
+    public List<ProductDTO> getUserProducts(String userId) {
+        List<Product> products = productRepository.findByUserId(userId);
+        return products.stream()
+                       .map(this::convertToDto)
+                       .collect(Collectors.toList());
     }
 
-    public Optional<Product> getProductById(String id) {
-        return productRepository.findById(id);
+    public Optional<ProductDTO> getProductById(String id, String userId) {
+        return productRepository.findById(id)
+                .filter(product -> product.getUserName().equals(userId))
+                .map(this::convertToDto);
     }
 
-    public Product saveProduct(Product product) {
+    public Product saveProduct(ProductDTO productDTO) {
+        Product product = convertToEntity(productDTO);
         return productRepository.save(product);
     }
 
-    public void deleteProduct(String id) {
-        productRepository.deleteById(id);
+    public void deleteProduct(String id, String userNsme) {
+        productRepository.findById(id).ifPresent(product -> {
+            if (product.getUserName().equals(userNsme)) {
+                productRepository.deleteById(id);
+            }
+        });
+    }
+
+    private ProductDTO convertToDto(Product product) {
+        return modelMapper.map(product, ProductDTO.class);
+    }
+
+    private Product convertToEntity(ProductDTO productDTO) {
+        return modelMapper.map(productDTO, Product.class);
     }
 }
